requirements.txt
keyboard>=0.13.5
pyserial>=3.5
